Split Tool API from CLI
Keep CLI entry points that can either be executed as console
scripts (e.g., `grammarinator-process`) or as modules (e.g.,
`python -m grammarinator.process`) in the `grammarinator` modules,
but move tool classes that can be invoked via the Tool API (e.g.,
`FuzzerFactory`) into a submodule called `grammarinator.tool`.

In addition to nicer separation of logic, this approach helps to
get rid of runtime warnings about import order causing potentially
unpredictable behaviour.

Also rename ANTLRv4 resource and parser directories.
