--- docs/grammarinator.rst
@@ -1,8 +1,0 @@
-===================================
-Tool API: package ``grammarinator``
-===================================
-
-.. automodule:: grammarinator
-   :members:
-   :imported-members:
-   :special-members: __call__, __exit__
