+++ docs/grammarinator.tool.rst
@@ -1,0 +1,8 @@
+========================================
+Tool API: package ``grammarinator.tool``
+========================================
+
+.. automodule:: grammarinator.tool
+   :members:
+   :imported-members:
+   :special-members: __call__, __exit__
