+++ grammarinator/tool/__init__.py
@@ -1,10 +1,10 @@
-# Copyright (c) 2017-2023 Renata Hodovan, Akos Kiss.
+# Copyright (c) 2023 Renata Hodovan, Akos Kiss.
 #
 # Licensed under the BSD 3-Clause License
 # <LICENSE.rst or https://opensource.org/licenses/BSD-3-Clause>.
 # This file may not be copied, modified, or distributed except
 # according to those terms.
 
-from . import runtime
-from . import tool
-from .pkgdata import __version__
+from .generate import Generator
+from .parse import ParserFactory
+from .process import FuzzerFactory
