+++ grammarinator/tool/resources/g4/ANTLRv4Lexer.g4
