+++ grammarinator/tool/resources/g4/LexBasic.g4
