+++ setup.cfg
@@ -43,6 +43,6 @@ console_scripts =
 
 [build_antlr]
 commands =
-    antlerinator:4.12.0 grammarinator/resources/antlr/ANTLRv4Lexer.g4 grammarinator/resources/antlr/ANTLRv4Parser.g4 -Dlanguage=Python3 -o grammarinator/parser -Xexact-output-dir -no-listener
+    antlerinator:4.12.0 grammarinator/tool/resources/g4/ANTLRv4Lexer.g4 grammarinator/tool/resources/g4/ANTLRv4Parser.g4 -Dlanguage=Python3 -o grammarinator/tool/g4 -Xexact-output-dir -no-listener
 output =
-    grammarinator/parser/ANTLRv4*.py
+    grammarinator/tool/parser/ANTLRv4*.py
