+++ tox.ini
@@ -13,7 +13,7 @@ commands = py.test --cov=grammarinator --cov-config=tox.ini {posargs}
 usedevelop = true
 
 [coverage:run]
-omit = **/parser/*
+omit = **/tool/g4/*
 
 [testenv:lint]
 deps =
@@ -21,7 +21,7 @@ deps =
     pylint
 commands =
     pylint grammarinator
-    pycodestyle grammarinator --ignore=E501 --exclude=grammarinator/parser/ANTLRv4*.py
+    pycodestyle grammarinator --ignore=E501 --exclude=grammarinator/tool/g4/ANTLRv4*.py
 
 [testenv:docs]
 deps = -rdocs/requirements.txt
